Fix compilation issues on other unixes
